+++ examples/common/BusDiag.h
@@ -1,11 +1,9 @@
 #pragma once
 
-#include <Wire.h>
-
 #include "I2cScanner.h"
 
 namespace bus_diag {
 inline void scan() {
-  i2c_scanner::scan(Wire);
+  i2c_scanner::scanDefault();
 }
 }