+++ examples/common/HealthDiag.h
@@ -10,7 +10,11 @@
 
 #pragma once
 
+#if defined(SSD1315_EXAMPLE_PLATFORM_IDF)
+#include "examples/common/IdfArduinoCompat.h"
+#else
 #include <Arduino.h>
+#endif
 
 #include "ssd1315/SSD1315.h"
 #include "ssd1315/Status.h"