+++ examples/common/HealthView.h
@@ -1,6 +1,10 @@
 #pragma once
 
+#if defined(SSD1315_EXAMPLE_PLATFORM_IDF)
+#include "examples/common/IdfArduinoCompat.h"
+#else
 #include <Arduino.h>
+#endif
 
 template <typename DriverT>
 inline void printHealthView(const DriverT& driver) {