+++ examples/common/I2cTransport.h
@@ -11,6 +11,9 @@
 
 #pragma once
 
+#if defined(SSD1315_EXAMPLE_PLATFORM_IDF)
+#include "examples/common/IdfI2cTransport.h"
+#else
 #include <Arduino.h>
 #include <Wire.h>
 
@@ -90,7 +93,9 @@ inline SSD1315::Status wireWrite(uint8_t addr, const uint8_t* data, size_t len,
  * @param timeoutMs I2C timeout in milliseconds (default 50ms)
  * @return true on success
  */
-inline bool initWire(int sda, int scl, uint32_t freq = 400000, uint16_t timeoutMs = 50) {
+inline bool initWire(int sda, int scl, uint32_t freq = 400000, uint16_t timeoutMs = 50,
+                     uint8_t address = 0x3C) {
+  (void)address;
   // First, try to recover the bus in case it's stuck from a previous crash
 #if defined(ARDUINO_ARCH_ESP32)
   // Toggle SCL to release any stuck slave
@@ -122,5 +127,11 @@ inline bool initWire(int sda, int scl, uint32_t freq = 400000, uint16_t timeoutM
   return true;
 }
 
+inline void* configUser() {
+  return &Wire;
+}
+
 }  // namespace transport
+
+#endif